requirements.txt
asgiref==3.2.5
dj-database-url==0.5.0
Django==3.0.4
django-ckeditor==5.9.0
django-js-asset==1.2.2
gunicorn==20.0.4
mysqlclient==1.4.6
psycopg2==2.8.4
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
